Refactor `files` handler
