+++ backend-go/internal/core/files/handler.go
@@ -1,17 +1,29 @@
 package files
 
 import (
+	"errors"
+	"fmt"
 	"net/http"
+	"strings"
 
 	"github.com/labstack/echo/v4"
+	"github.com/lithammer/shortuuid/v4"
 
 	"github.com/yorkie-team/codepair/backend/api/codepair/v1/models"
+	"github.com/yorkie-team/codepair/backend/internal/infra/database"
+	"github.com/yorkie-team/codepair/backend/internal/infra/database/entity"
+	"github.com/yorkie-team/codepair/backend/internal/infra/storage"
 	"github.com/yorkie-team/codepair/backend/internal/middleware"
+)
+
+const (
+	maxContentLength = 10_000_000
 )
 
 // Handler handles HTTP requests for files services.
 type Handler struct {
-	service *Service
+	storageClient storage.Client
+	workspaceRepo Repository
 }
 
 // createUploadPresignedURL handles POST /files requests.
@@ -25,17 +37,40 @@ func (h *Handler) createUploadPresignedURL(c echo.Context) error {
 		return middleware.NewError(http.StatusBadRequest, "validation failed", err)
 	}
 
-	resp, err := h.service.createUploadPresignedURL(
+	workspace, err := h.workspaceRepo.FindWorkspaceByID(entity.ID(req.WorkspaceId))
+	if err != nil {
+		if errors.Is(err, database.ErrWorkspaceNotFound) {
+			return WorkspaceNotFoundError
+		}
+
+		return middleware.NewError(http.StatusInternalServerError, "server internal error", err)
+	}
+
+	if req.ContentLength > maxContentLength {
+		return ContentLengthTooLongError
+	}
+
+	extension := ""
+	if parts := strings.Split(req.ContentType, "/"); len(parts) > 1 {
+		extension = parts[1]
+	}
+
+	fileKey := generateFileKey(workspace.Slug, extension)
+
+	url, err := h.storageClient.CreateUploadPresignedURL(
 		c.Request().Context(),
-		req.WorkspaceId,
+		fileKey,
 		req.ContentLength,
 		req.ContentType,
 	)
 	if err != nil {
-		return err
+		return middleware.NewError(http.StatusInternalServerError, "server internal error", err)
 	}
 
-	return c.JSON(http.StatusOK, resp)
+	return c.JSON(http.StatusOK, &models.CreateUploadPresignedUrlResponse{
+		FileKey: fileKey,
+		Url:     url,
+	})
 }
 
 // createDownloadPresignedURL handles GET /files/:file_name requests.
@@ -45,10 +80,15 @@ func (h *Handler) createDownloadPresignedURL(c echo.Context) error {
 		return middleware.NewError(http.StatusBadRequest, "file_name is required")
 	}
 
-	url, err := h.service.createDownloadPresignedURL(c.Request().Context(), fileKey)
+	url, err := h.storageClient.CreateDownloadPresignedURL(c.Request().Context(), fileKey)
 	if err != nil {
-		return err
+		return FileNotFoundError
 	}
 
 	return c.Redirect(http.StatusFound, url)
+}
+
+// generateFileKey generates a file key with slug and extension.
+func generateFileKey(slug, extension string) string {
+	return fmt.Sprintf("%s-%s.%s", slug, shortuuid.New(), extension)
 }
